Implement early initialization
Change the Epic Games SDK to initialize the SDK in the wrapper extension Init() method, which is called before the web content is loaded, rather than in an "init" message sent from the Construct plugin, which is sent after the web content has loaded. This is in an effort to make sure the SDK can hook in to anything that happens during loading of the web content, which may help with integration. To access the plugin properties, there is a new system to export plugin properties to package.json, where the wrapper extension can then parse the details it needs. NOTE: this relies on features that are currently in development and yet to be released - wait until at least the next beta release before testing this.
Note this commit also adjusts logging so messages are logged both to the debug output and also the browser console with the LogToConsole() method, so the activity and status of the wrapper extension can more easily be observed in the usual browser dev tools. It also simplifies some of the callback code with lambdas.